explored options to include NOT into joins
This is an abandoned attempt to fix the problem than
NOT applied to a field on a sub document always gets
translated to NOT ANY. After testing several approaches
I came to the conclusion that this is a principal issue
with the negation and boolean logic.
When we allow a NOT to become an inlined condition
(on a particular sub document rather than any) we open
Pandora's box because the same
profile may appear in the set with the positive condition
C AND also in the set NOT(C) (because a different sub-document
instance matches the NOT). This feels so fundamentally wrong and
cannot be explained to any user that it is better to have a
known framework limitation that NOT always means NOT ANY.